+++ src/features/manageProfileSettings/index.ts
@@ -3,6 +3,7 @@ export { DefaultProfileOption } from './components/DefaultProfileOption';
 export { ManageProfileSettingsControlPane } from './components/ManageProfileSettingsControlPane';
 export { ManageProfileSettingsControls } from './components/ManageProfileSettingsControls';
 export { ModalCloseProfileSettings } from './components/ModalCloseProfileSettings';
+export { ModalSaveUnusedProfileComponents } from './components/ModalSaveUnusedProfileComponents';
 export { ProfileSettings } from './components/ProfileSettings';
 export { ProfilesList } from './components/ProfilesList';
 
@@ -16,11 +17,14 @@ export {
   useProfileList,
   usePreferredProfiles,
   useSettingsAnnouncements,
+  useSaveProfileSettings,
 } from './hooks';
 export {
   childrenDifference,
   chooseModifiers,
   componentFromId,
+  determinePreferredAction,
+  generateSettings,
   getProfileChildren,
   getSettingsChildren,
   FilteredKeyboardSensor,