+++ src/features/manageProfileSettings/utils/index.ts
@@ -1,3 +1,4 @@
 export { FilteredPointerSensor, FilteredKeyboardSensor } from './filteredSensors';
 export { componentFromId, getProfileChildren, getSettingsChildren, childrenDifference } from './children';
 export { chooseModifiers } from './modifiers';
+export { determinePreferredAction, generateSettings } from './profileSave';