+++ CHANGELOG.md
@@ -85,6 +85,7 @@
 * Add mandatory field handling in custom settings. Refs [UILD-728].
 * Add possibility to assign Hubs as Subjects. Refs [UILD-711].
 * Add possibility to change the assigned Hubs as Subjects. Refs [UILD-730].
+* Connect profile settings editor to save settings API. Refs [UILD-698].
 
 [UILD-552]:https://folio-org.atlassian.net/browse/UILD-552
 [UILD-544]:https://folio-org.atlassian.net/browse/UILD-544
@@ -169,6 +170,7 @@
 [UILD-728]:https://folio-org.atlassian.net/browse/UILD-728
 [UILD-711]:https://folio-org.atlassian.net/browse/UILD-711
 [UILD-730]:https://folio-org.atlassian.net/browse/UILD-730
+[UILD-698]:https://folio-org.atlassian.net/browse/UILD-698
 
 ## 1.0.5 (2025-04-30)
 * Browser Back button navigates to Edit without duplicated title when navigated from Duplicate screen. Fixes [UILD-554].